requirements.txt
folium==0.13.0
haversine==2.7.0
numpy==1.23.1
pandas==1.4.3
plotly==5.10.0
streamlit==1.15.1
streamlit-folium==0.7.0
Pillow==9.2.0
matplotlib-inline==0.1.6
matplotlib=3.5.3